project/requirements.txt
aioboto3==8.3.0
aiobotocore==1.2.2
aiofiles==0.6.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
boto3==1.16.52
botocore==1.19.52
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.3
django-redis==5.0.0
djangorestframework==3.12.4
entrypoints==0.3
future==0.18.2
graphviz==0.16
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient @ file:///D:/Web%E5%BC%80%E5%8F%91/Web%E6%A1%86%E6%9E%B6/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.2
objgraph==3.5.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==1.1.1
pyzmq==22.1.0
redis==3.5.3
requests==2.25.1
s3transfer==0.3.7
Send2Trash==1.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.10.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
views==0.3
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
